+++ src/Parameters.php
@@ -59,7 +59,7 @@ final class Parameters implements MemberOrderedMap
     /**
      * Returns a new instance.
      */
-    public static function create(): self
+    public static function new(): self
     {
         return new self();
     }