+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - `Item::fromString` to improve Item public API;
 - `Value` internal class to improve Item public API;
 - `Token::toString` to return the string representation of the token.
+- `Parameters::new` and `Dictionary::new` to return a new and empty instance
 
 ### Fixed
 
@@ -32,6 +33,8 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - **[BC Break]** `Value` interface use a combination of `ValueAccess` **and** `ParameterAccess` instead.
 - **[BC Break]** `Token::value` is no longer public use `Token::toString` instead.
 - **[BC Break]** `Item::from` is removed use `Item::fromAssociative` instead.
+- **[BC Break]** `Parameters::create` is removed use `Parameters::new` instead.
+- **[BC Break]** `Dictionary::create` is removed use `Dictionary::new` instead.
 
 ## [0.8.0] - 2023-03-12